requirements.txt
fastapi
langchain
langchain-google-genai
langchain_core
langchain_community
uvicorn[standard]
python-dotenv
python-pptx
unstructured 
youtube-transcript-api 
docx2txt 
networkx 
pandas 
xlrd 
openpyxl 
gdown 
pytest 
PyPDF2 
python-dotenv 
psutil
pytube
gradio

google-cloud-secret-manager
google-cloud-logging
google-auth
google-cloud-storage
firebase-admin